@@ -1,15 +1,12 @@
 import { StrictMode } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-  document.getElementById('root')
+const root = createRoot(document.getElementById('root') as HTMLElement);
+
+root.render(
+	<StrictMode>
+		<App />
+	</StrictMode>
 );
-
-
-reportWebVitals();
